+++ Calculadora/Pila.h
@@ -1,8 +1,8 @@
 // Todas las operaciones de la Pila.c
-void push(int a);
+void push(double a);
 void pop();
 void show();
 void wipe();
 int size();
 int length();
-int lastelement();
+double lastelement();
